requirements.txt
# AI és Workflow Management (optimalizált)
langgraph>=0.2.74
pydantic-ai-slim[openai,anthropic,logfire]>=0.0.49
langchain-core>=0.3.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
# openai és anthropic automatikusan települ a langchain csomagokkal

# Web Framework és API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0

# Security és Rate Limiting
slowapi>=0.1.9
python-multipart>=0.0.6
fastapi-csrf-protect>=0.3.2
# Additional security dependencies
cryptography>=41.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# JWT és token kezelés
PyJWT>=2.8.0
# Input validation és sanitization
bleach>=6.0.0
# Security headers
secure>=0.3.0
# Rate limiting
limits>=3.6.0
# IP filtering
ipaddress>=1.0.23
# LangGraph SDK (hivatalos authentication)
langgraph-sdk>=0.0.1
# Security monitoring
logfire>=0.0.1
# Environment validation
pydantic-settings>=2.0.0
# Additional security tools
bandit>=1.7.0  # Security linter
safety>=2.3.0  # Dependency vulnerability scanner

# Adatbázis és Cache
supabase>=2.3.0
asyncpg>=0.29.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Vector Database (pgvector)
pgvector>=0.2.5
sentence-transformers>=2.2.0
numpy>=1.24.0
# OpenAI embeddings
openai>=1.0.0
# Vector similarity search
scikit-learn>=1.3.0

# Marketing Automation & Messaging
sendgrid>=6.9.0
twilio>=8.2.0
jinja2>=3.1.0

# Social Media Integrations (egyszerűsített)
requests>=2.31.0
# facebook-sdk és whatsapp-business-api helyett requests használata

# Background Task Processing
celery>=5.3.0
kombu>=5.3.0

# Email Templates
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Monitoring és Logging
structlog>=23.1.0
pydantic-logfire==0.0.1  # Python 3.13 kompatibilitás

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0